backend/requirements.txt
alembic
faker
fastapi
jellyfish
pandas
psycopg2-binary
pydantic[email]
python-dotenv
python-multipart
rapidfuzz
ruff
sqlalchemy
uvicorn[standard]
